Remove private IP filtering
I'm thinking of implementing blocking using source-destination IP pairs.
Or better yet, using transport layer ports as well to identify
individual flows. If I manage to do that, there won't be need to
check for private IP addresses because we won't inadvertently break all
communication of an IP address; we will only be blocking on a per-flow basis.