Add table entries to the switch based on capture results
Tested in the example topology; seems to work.
I was able to identify and block ICMP and SSH successfully.

Uses the "management interface" that I defined to connect to the switch
from the monitoring device. The IP you will see in blocklist_add.py
corresponds to the IP of that interface.

Most important changes:
- Moved the switch connection part to a separate Python 2 program.
  The main program runs this as a subprocess to add entries to the table.
  This way I avoided to port the main program to Python 2 (remember that
  p4runtime_lib works with Python 2, which is what made me consider making
  a mess like this in the first place).
- Added another required parameter: --bmv2_json. The JSON representation
  of the P4 program, generated by p4c. Used by the ./blocklist_add.py
  program. Apparently we need to make a call to SetForwardingPipelineConfig()
  in order to set table entries. That call requires the JSON file.

The code may need a little refactoring before making further extensions
or things will probably start falling apart. It is a bunch of hacks here
and there that I made to get things working as quickly as possible.